requirements.txt
pandas==2.2.2
pymupdf==1.24.6
openai==1.35.5
python-dotenv==1.0.1
transformers==4.42.3
torch==2.4.0
langchain==0.2.6
openpyxl==3.1.5
numpy==1.26.4
gensim==4.3.2
nltk==3.8.1
scipy==1.14.0
POT==0.9.4
matplotlib==3.9.1
seaborn==0.13.2
tqdm==4.66.4
tiktoken==0.7.0
sentencepiece==0.2.0
streamlit==1.38.0
streamlit-pdf-viewer==0.0.18rc2
langchain-community==0.2.16
PyPDF2==3.0.1
faiss-cpu==1.8.0